code clean up
-deleted unnecessary js files and test html files from repository
-added a template and a new wrapper object
-CSS
--Added a few styles for the map pins' info window
-HTML
--Added Data Driven Documents(D3) library
--Added script files for query wrapper objects
--changed some class names and changed greater than checkboxes to
between some range checkboxes
--top level element of the templates now use dbn numbers for class
names.
-JS
--created parseWrapper to handle queries, simplify OR queries and add
AND queries. The promise is for an array of only attributes. The JSON
needed without all of the nonsense parse.com queries come with
--SODA also returns just the JSON needed and all the query functions
have an extra parameter just because browser might complain about
insufficient parameters. $.getJSON().then() save an array with the JSON
and the status because the callback function only had one variable
instead of 2. This lead to this silly line: searchResults =
searchResults[0];
--makeMarker() does just that and attaches a click event listener that
will run popPin().
--popPin opens up an infoWindow on the pin that was clicked.
--clicking a search result's name does not fade the map anymore. Instead
it does something more useful. it finds a corresponding pin and calls
popPin(). See above for current details.
--added copy() to Parse.Query object. parseWrapper needs this to make
AND possible
--edited refreshResults() to adapt to the new wrapper additions. The
code is a lot cleaner and much less redundant. Added functions to handle
repetitive building query logic. No more need for
parseResult.attributes. Map is now centered in the center of all the
points.
--Added clearAll() to myFutureHS to be called from refreshResults().
Added some more properties to myFutureHS for demographics and
google.maps.infoWindow. Added event listener to close the infoWindow
when the map is clicked on an empty spot.